# backend/openapi/unital-v1.yaml
openapi: 3.1.0
info:
  title: Unital API
  version: 1.0.0
  description: |
    Building & complex management platform. Consolidated gateway API for the
    Unital Go microservices (identity, property, billing, facilities,
    operations, notifications/communications).

    **Auth**: `POST /auth/login` returns an access token (15 min) + refresh
    token. Send `Authorization: Bearer <access>` on every other call.

    **Errors**: all failures use RFC 9457 problem details with a
    machine-readable `code` and `trace_id`.
  contact: { name: Unital, url: https://unital.app }
servers:
  - url: http://localhost:8080/api/v1
    description: local gateway
  - url: http://localhost:18080/api/v1
    description: local gateway (dev port)
...
security:
  - bearerAuth: []
paths:
...
  /templates:
    get:
      tags: [Notifications]
      summary: List notification templates
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/NotificationTemplate" }
    post:
      tags: [Notifications]
      summary: Create/update a template (per-channel variants)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationTemplate" }
      responses:
        "200": { description: Upserted }
  /notifications:send:
    post:
      tags: [Notifications]
      summary: Send a campaign (fan-out to recipients × channels)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [building_id, template, target]
              properties:
                building_id: { type: string }
                template: { type: string, example: charge.overdue.reminder }
                target: { $ref: "#/components/schemas/NotificationTarget" }
                vars:
                  type: object
                  additionalProperties: { type: string }
                  example: { period: "1405-06", remaining: "1500000" }
                channels:
                  type: array
                  items: { type: string, enum: [inapp, email, sms, webhook] }
                schedule_at: { type: string, format: date-time }
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /notifications/{campaignID}:
    get:
      tags: [Notifications]
      summary: Campaign status + per-channel counters + deliveries
      parameters:
        - name: campaignID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaign: { $ref: "#/components/schemas/Campaign" }
                  counters:
                    type: object
                    properties:
                      pending: { type: integer }
                      sent: { type: integer }
                      failed: { type: integer }
                  deliveries:
                    type: array
                    items: { $ref: "#/components/schemas/Delivery" }
  /me/notifications:
    get:
      tags: [Notifications]
      summary: My inbox
      parameters:
        - { name: unread, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Inbox messages (newest first)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/InboxMessage" }
  /me/notifications/{messageID}/read:
    post:
      tags: [Notifications]
      summary: Mark an inbox message read
      parameters:
        - name: messageID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Marked read }
...
  /buildings/{buildingID}/announcements:
    parameters:
      - $ref: "#/components/parameters/BuildingID"
    get:
      tags: [Comms]
      summary: List announcements (members see published; managers see drafts too)
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Announcement" }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      tags: [Comms]
      summary: Draft an announcement (manager/board)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnouncementInput" }
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "403": { $ref: "#/components/responses/Problem" }
  /announcements/{announcementID}:
    parameters:
      - name: announcementID
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Comms]
      summary: Announcement detail
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [Comms]
      summary: Edit a draft announcement (manager/board; published ones immutable)
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnouncementInput" }
      responses:
        "200": { description: Updated }
        "422": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [Comms]
      summary: Delete an announcement (manager/board)
      responses:
        "200": { description: Deleted }
        "403": { $ref: "#/components/responses/Problem" }
  /announcements/{announcementID}/publish:
    post:
      tags: [Comms]
      summary: Publish and fan out to the targeted audience's in-app inbox
      parameters:
        - name: announcementID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Published (delivered_to counts inbox recipients)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Announcement" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /buildings/{buildingID}/meetings:
    parameters:
      - $ref: "#/components/parameters/BuildingID"
    get:
      tags: [Comms]
      summary: List meetings (staff see all; members see general/emergency)
      responses:
        "200":
          description: Meetings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Meeting" }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      tags: [Comms]
      summary: Schedule a meeting (manager/board)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MeetingInput" }
      responses:
        "201":
          description: Scheduled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meeting" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /meetings/{meetingID}:
    parameters:
      - name: meetingID
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Comms]
      summary: Meeting detail
      responses:
        "200":
          description: Meeting
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meeting" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [Comms]
      summary: Update a meeting (manager/board; status transitions live here)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                agenda: { type: string }
                location: { type: string }
                status: { type: string, enum: [scheduled, ongoing, completed, cancelled] }
      responses:
        "200": { description: Updated }
        "403": { $ref: "#/components/responses/Problem" }
  /meetings/{meetingID}/rsvp:
    post:
      tags: [Comms]
      summary: RSVP (any building member; upserts the latest status)
      parameters:
        - name: meetingID
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [confirmed, declined, attended, absent] }
      responses:
        "200":
          description: Attendance recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendance" }
        "422": { $ref: "#/components/responses/Problem" }
  /meetings/{meetingID}/attendance:
    get:
      tags: [Comms]
      summary: Attendance roster (manager/board)
      parameters:
        - name: meetingID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Attendance rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Attendance" }
        "403": { $ref: "#/components/responses/Problem" }
  /meetings/{meetingID}/minutes:
    parameters:
      - name: meetingID
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Comms]
      summary: Read the minutes
      responses:
        "200":
          description: Minutes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Minute" }
        "404": { $ref: "#/components/responses/Problem" }
    put:
      tags: [Comms]
      summary: Write/update the minutes (manager/board)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Minute" }
      responses:
        "200": { description: Saved }
        "403": { $ref: "#/components/responses/Problem" }
  /meetings/{meetingID}/minutes/sign:
    post:
      tags: [Comms]
      summary: Sign the minutes as the current user (idempotent)
      parameters:
        - name: meetingID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Signed minutes }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /buildings/{buildingID}/tickets:
    parameters:
      - $ref: "#/components/parameters/BuildingID"
    get:
      tags: [Comms]
      summary: List support tickets (staff see all; residents their own)
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Ticket" }
    post:
      tags: [Comms]
      summary: Open a support ticket (any building member)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TicketInput" }
      responses:
        "201":
          description: Ticket opened
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /tickets/{ticketID}:
    get:
      tags: [Comms]
      summary: Ticket detail with thread (staff or submitter)
      parameters:
        - name: ticketID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Ticket + responses
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket: { $ref: "#/components/schemas/Ticket" }
                  responses:
                    type: array
                    items: { $ref: "#/components/schemas/TicketResponse" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
  /tickets/{ticketID}/close:
    post:
      tags: [Comms]
      summary: Close a ticket (staff)
      parameters:
        - name: ticketID
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Closed }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Problem:
      description: Error (RFC 9457)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    NotificationTemplate:
      type: object
      required: [name, variants]
      properties:
        name: { type: string, example: charge.overdue.reminder }
        severity: { type: string, enum: [normal, urgent] }
        channels:
          type: array
          items: { type: string, enum: [inapp, email, sms, webhook] }
        variants:
          type: object
          additionalProperties:
            type: object
            properties:
              title: { type: string }
              body: { type: string, description: "Supports {{var}} placeholders" }
...